+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2021-05-24
+
+- Add SingleSelect and MultiSelect components
+- Removed deprecated Select and ComboMulti
+
 ## [1.2.0] - 2021-05-21
 
 - Add eslint with CRA rules